# swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "title": "github.com/Phi-S/cs-server-manager API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/settings": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Get the current settings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SettingsModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Update settings",
                "parameters": [
                    {
                        "description": "The updated settings",
                        "name": "settings",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.SettingsModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SettingsModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/start": {
            "post": {
                "description": "Starts the server with the given start parameters",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "server"
                ],
                "summary": "Start the server",
                "parameters": [
                    {
                        "description": "You can provide no, all or only a few start parameters. The provided start parameters will overwrite the saved start parameters in the start-parameters.json file if the server started successfully.",
                        "name": "startParameters",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/handlers.StartBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/stop": {
            "post": {
                "description": "Stops the server of if the server is not running, returns 200 OK",
                "tags": [
                    "server"
                ],
                "summary": "Stop the server",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.SettingsModel": {
            "type": "object",
            "required": [
                "hostname",
                "max_players",
                "start_map"
            ],
            "properties": {
                "hostname": {
                    "type": "string",
                    "maxLength": 128
                },
                "max_players": {
                    "type": "integer",
                    "maximum": 128
                },
                "password": {
                    "type": "string",
                    "maxLength": 32
                },
                "start_map": {
                    "type": "string",
                    "maxLength": 32
                },
                "steam_login_token": {
                    "type": "string"
                }
            }
        },
        "handlers.StartBody": {
            "type": "object",
            "properties": {
                "hostname": {
                    "type": "string",
                    "maxLength": 128
                },
                "max_players": {
                    "type": "integer",
                    "maximum": 128
                },
                "password": {
                    "type": "string",
                    "maxLength": 32
                },
                "start_map": {
                    "type": "string",
                    "maxLength": 32
                },
                "steam_login_token": {
                    "type": "string"
                }
            }
        },
...
    }
}